% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_tsheatmap.R
\name{geom_tsheatmap}
\alias{geom_tsheatmap}
\title{Add a layer of heatmap with a tip data using akima interpolation}
\usage{
geom_tsheatmap(trait, resolution = 0.001, bins = 24, ...)
}
\arguments{
\item{trait}{column name of the trait}

\item{resolution}{resolution of the heatmap}

\item{bins}{number of contour bins.}

\item{...}{additional parameters}
}
\value{
ggplot object
}
\description{
Add a layer of heatmap with a tip data using akima interpolation
}
\examples{
library(ggtree)
library(phytools)
tr <- rtree(15)
td <- fastBM(tr, nsim = 2, bounds = c(0, Inf))
tdex <- data.frame(z = fastBM(tr, nsim = 1),
                  node = 1:15)
p <- ggtreeSpace(tr, td)
p \%<+\% tdex +
 geom_tippoint() +
 geom_tsheatmap(trait = "z", alpha = 0.7 ,resolution = 0.01, bin = 12)
}
