% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggphylospm.R
\name{ggphylopsm}
\alias{ggphylopsm}
\title{Plot phylogenetic scatterplot matrix}
\usage{
ggphylopsm(
  tr,
  traits = NULL,
  title = "Phylogenetic Scatterplot Matrix",
  xAxisLabels = NULL,
  yAxisLabels = NULL,
  tr.params = list(size = 1, colors = NULL),
  sptr.params = list(tippoint = TRUE, tiplab = FALSE, labdir = "horizon", panel.grid =
    FALSE)
)
}
\arguments{
\item{tr}{}

\item{traits}{}

\item{title}{}

\item{xAxisLabels}{}

\item{yAxisLabels}{}

\item{tr.params}{}

\item{sptr.params}{}
}
\description{
Plot phylogenetic scatterplot matrix
}
