% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_treespace.R
\name{intern_call}
\alias{intern_call}
\title{Make plot data for ggtreespace.
This function processes a phylogenetic tree and associated trait data to
create a data frame suitable for plotting with `ggtreespace`.}
\usage{
intern_call(data, t)
}
\arguments{
\item{data}{Input data of check.graph.layout}

\item{t}{Input traits}
}
\value{
trait data.frame
}
\description{
Make plot data for ggtreespace.
This function processes a phylogenetic tree and associated trait data to
create a data frame suitable for plotting with `ggtreespace`.
}
\examples{
library(ggplot2)
library(ggtree)
library(phytools)

tr <- rtree(10)
a <- fastBM(tr, nsim = 2)

trd <- make_ts_data(tr, a)
}
