% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-ggplot-add.r
\docType{import}
\name{reexports}
\alias{reexports}
\alias{ggplot_add}
\title{Objects exported from other packages}
\value{
ggplot object
}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{ggplot2}{\code{\link[ggplot2:update_ggplot]{ggplot_add}}}
}}

