% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_treeSpace}
\alias{theme_treeSpace}
\title{Phylomorphospace theme No.1, with arrow-end axis and grey panel grid}
\usage{
theme_treeSpace(...)
}
\arguments{
\item{...}{additional parameters for theme()}
}
\value{
a theme object with arrow-end axis
}
\description{
Phylomorphospace theme No.1, with arrow-end axis and grey panel grid
}
\examples{
library(ggtree)
library(phytools)

tr <- rtree(15)
td <- fastBM(tr, nsim = 2)
ggtreeSpace(tr, td) +
 geom_tippoint() + 
 theme_treeSpace()
}
