% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_treeSpace2}
\alias{theme_treeSpace2}
\title{Phylomorphospace theme No.1,with blank background and panel border}
\usage{
theme_treeSpace2(...)
}
\arguments{
\item{...}{additional parameters for theme()}
}
\value{
a theme object with blank background and panel border
}
\description{
Phylomorphospace theme No.1,with blank background and panel border
}
\examples{
library(ggtree)
library(phytools)

tr <- rtree(15)
td <- fastBM(tr, nsim = 2)
ggtreeSpace(tr, td) +
 geom_tippoint() + 
 theme_treeSpace2()
}
