% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_treeSpace.R
\name{geom_treeSpace}
\alias{geom_treeSpace}
\title{A layer of phylomorphospace}
\usage{
geom_treeSpace(tr, tipdata, mapping = NULL, ...)
}
\arguments{
\item{tr}{a phylogenetic tree}

\item{tipdata}{tipdata}

\item{mapping}{mapping}

\item{...}{additional parameters}
}
\value{
ggplot object
}
\description{
A layer of phylomorphospace
}
\examples{
library(ggplot2)
library(ggtree)
library(phytools)

tr <- rtree(10)
a <- fastBM(tr, nsim = 2)

p <- ggplot() + geom_treeSpace(tr, a)
}
