% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_tsheatmap.R
\name{geom_tsheatmap}
\alias{geom_tsheatmap}
\title{Add a layer of heatmap with a tip data using akima interpolation}
\usage{
geom_tsheatmap(trait, resolution = 0.001, bins = 24, show.legend = FALSE, ...)
}
\arguments{
\item{trait}{column name of the trait}

\item{resolution}{resolution of the heatmap}

\item{bins}{Number of contour bins. Overridden by breaks.}

\item{show.legend}{}

\item{...}{}
}
\description{
Add a layer of heatmap with a tip data using akima interpolation
}
