% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggtreeSpace.R
\name{ggtreeSpace}
\alias{ggtreeSpace}
\title{Plot phylomorphospace}
\usage{
ggtreeSpace(tr, tipdata, mapping = NULL, ...)
}
\arguments{
\item{tr}{tree}

\item{tipdata}{tipdata containig tip coordinates}

\item{mapping}{aesthetic mapping}

\item{...}{additional parameters}
}
\value{
ggtreeSpace object
}
\description{
This function plot phylomorphospace
}
\examples{
library(ggtree)
library(ggplot2)
library(phytools)
tr <- rtree(15)
td <- fastBM(tr, nsim = 2)
ggtreeSpace(tr, td) +
 geom_tippoint()

}
