% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_treeSpace}
\alias{theme_treeSpace}
\title{Title}
\usage{
theme_treeSpace(...)
}
\arguments{
\item{...}{additional parameters for theme()}
}
\value{
updated ggplot object with new theme
}
\description{
Title
}
\examples{
library(ggtree)
library(phytools)

tr <- rtree(15)
td <- fastBM(tr, nsim = 2)
ggtreeSpace(tr, td) +
 geom_tippoint() + 
 theme_treeSpace()
}
