% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_treespace2}
\alias{theme_treespace2}
\title{Phylomorphospace theme No.2, with blank background and panel border.}
\usage{
theme_treespace2(...)
}
\arguments{
\item{...}{additional parameters for `ggplot2::theme`}
}
\value{
a theme object with blank background and panel border
}
\description{
Phylomorphospace theme No.2, with blank background and panel border.
}
\examples{
library(ggtree)
library(phytools)
library(ggtreeSpace)

tr <- rtree(15)
td <- fastBM(tr, nsim = 2)
ggtreespace(tr, td) +
 geom_tippoint() + 
 theme_treespace2()
}
